Implemented Generate Report feature
The total cost has to be revised to actually print the total price of the service provided, and change the html into a stringview, i didnt want to change the layout myself because i was scared the webview and the button would go crazy. Otherwise, the feature works. The file gets created successfully!